requirements.txt
fastapi
asyncio
python-dotenv
pydantic
requests
redis
mysql
uvicorn